+++ elastic-job-core/src/main/java/com/dangdang/ddframe/job/api/ElasticJob.java
@@ -4,9 +4,9 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
- *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -17,37 +17,10 @@
 
 package com.dangdang.ddframe.job.api;
 
-import com.dangdang.ddframe.job.exception.JobException;
-import com.dangdang.ddframe.job.internal.schedule.JobFacade;
-
 /**
- * 弹性化分布式作业接口.
+ * 弹性化分布式作业标识接口.
  * 
  * @author zhangliang
  */
 public interface ElasticJob {
-    
-    /**
-     * 执行作业.
-     */
-    void execute();
-    
-    /**
-     * 处理作业执行时异常.
-     * 
-     * @param jobException 作业异常
-     */
-    void handleJobExecutionException(JobException jobException);
-    
-    /**
-     * 获取提供作业服务的门面类.
-     */
-    JobFacade getJobFacade();
-    
-    /**
-     * 设置提供作业服务的门面类.
-     * 
-     * @param jobFacade 提供作业服务的门面类
-     */
-    void setJobFacade(JobFacade jobFacade);
 }